pyproject.toml
[tool.poetry]
name = "baza-backend"
version = "0.1.0"
description = "API backend for Baza Foundation"
authors = ["Prasanta <prasantakakati1994@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
django = "2.1.5"
django-rest-framework = "^0.1.0"
django-cors-headers = "2.4.0"
uwsgi = "2.0.17.1"
"psycopg2-binary" = "2.8.5"
raven = "6.10.0"
channels = "2.1.6"
channels-redis = "2.3.2"
django-redis-cache = "2.1.0"
pillow = "5.4.0"
celery = "4.4.0"
twilio = "6.22.1"
django-simple-history = "2.6.0"
"iso3166" = "0.9"
geopy = "1.18.1"
"geoip2" = "2.9.0"
pyotp = "2.2.7"
oauthlib = "2.1.0"
django-oauth-toolkit = "1.2.0"
django-versatileimagefield = "1.10"
bleach = "3.0.2"
service-identity = "18.1.0"
coinbase-commerce = "1.0.1"
django-mjml = "0.6.0"
tweepy = "3.7.0"
facepy = "1.0.9"
django-extra-fields = "1.0.0"
markdown = "^3.1"
beautifulsoup4 = "^4.8"

[tool.poetry.dev-dependencies]
"pep8" = "1.7.1"
"autopep8" = "1.4"
ipython = "6.5.0"
"flake8" = "3.5.0"
python-language-server = {version = "*", extras = ["all"]}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
